python/url-shorter-django/requirements.txt
Django==2.2
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
requests==2.28.1
